refactor: extract bot detection logic into a reusable DetectsBots trait and integrate it into middleware
