+++ app/Http/Middleware/LogSiteVisitor.php
@@ -8,6 +8,8 @@ use Symfony\Component\HttpFoundation\Response;
 
 class LogSiteVisitor
 {
+    use \App\Traits\DetectsBots;
+
     /**
      * Handle an incoming request.
      *
@@ -55,30 +57,6 @@ class LogSiteVisitor
         return $response;
     }
 
-    /**
-     * Determine if the user agent belongs to a bot.
-     */
-    protected function isBot(string $userAgent): bool
-    {
-        if (empty($userAgent)) {
-            return true; // Không có user-agent thường là bot/tool
-        }
-
-        $bots = [
-            'bot', 'crawl', 'spider', 'slurp', 'google', 'bing', 'yandex', 'duckduck', 'baidu',
-            'ahrefs', 'semrush', 'dotbot', 'rogerbot', 'exabot', 'mj12bot', 'archive', 'pinger',
-            'screaming', 'headless', 'inspect', 'lighthouse', 'python', 'curl', 'wget', 'php',
-            'java', 'perl', 'ruby', 'go-http', 'urllib', 'httpclient', 'scrapy'
-        ];
-
-        $userAgent = strtolower($userAgent);
-
-        foreach ($bots as $bot) {
-            if (str_contains($userAgent, $bot)) {
-                return true;
-            }
-        }
-
-        return false;
-    }
+    // Removed local isBot method as it's now in DetectsBots trait
 }
+
